Removal of majority of remaining unused variables (with exception of some ionisation routines that will be addressed in a future commit).  Also cleaned up some code, indents, removing tabs, etc..

+++ src/Headers/NeighbourManager.h
@@ -342,7 +342,7 @@ public:
    {
     TrimNeighbourLists<InParticleType,_false_type>(p, hydromask, true);
 
-    assert((culled_neiblist.size()+directlist.size()+smoothgravlist.size()) == GetNumAllNeib());
+    assert((int) (culled_neiblist.size()+directlist.size()+smoothgravlist.size()) == GetNumAllNeib());
 
 
     typedef typename GravityNeighbourLists<ParticleType>::DirectType DirectType ;
